server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
